+++ src/single_thread.rs
@@ -1,6 +1,6 @@
 use std::sync::{Arc,Mutex};
 
-pub fn mergesort(arr: Arc<Mutex<Vec<isize>>>, begin: usize, end: usize) {
+pub fn mergesort(arr: Arc<Mutex<Vec<i16>>>, begin: usize, end: usize) {
     if end - 1 > begin {
         let mid = begin + (end - begin) / 2;
         mergesort(Arc::clone(&arr), begin, mid);
@@ -9,7 +9,7 @@ pub fn mergesort(arr: Arc<Mutex<Vec<isize>>>, begin: usize, end: usize) {
     }
 }
 
-fn merge(arr: Arc<Mutex<Vec<isize>>>, begin: usize, mid: usize, end: usize) {
+fn merge(arr: Arc<Mutex<Vec<i16>>>, begin: usize, mid: usize, end: usize) {
     let left_size = mid - begin;
     let right_size = end - mid;
     let mut numbers = arr.lock().unwrap();
